[Refactor] Move type into new file since it didn't belong
