+++ modules/AlphaWalletFoundation/AlphaWalletFoundation/KeyManagement/SignMessageType.swift
@@ -1,0 +1,10 @@
+// Copyright © 2023 Stormbird PTE. LTD.
+
+import Foundation
+
+public enum SignMessageType {
+    case message(Data)
+    case personalMessage(Data)
+    case typedMessage([EthTypedData])
+    case eip712v3And4(EIP712TypedData)
+}
